safari-600.7-branch/LayoutTests/userscripts/user-script-plugin-document-expected.txt

CONSOLE MESSAGE: line 1: End Script: user-script-plugin-document.html
CONSOLE MESSAGE: line 1: Start Script: data:application/x-webkit-test-netscape,
CONSOLE MESSAGE: line 1: End Script: data:application/x-webkit-test-netscape,
